requirements.txt
asgiref==3.5.0
Django==4.0.1
django-tinymce==3.4.0
gunicorn==20.1.0
Pillow==9.0.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
